+++ pyincore/analyses/epfdamage/test_epfdamage.py
@@ -16,10 +16,10 @@ def run_with_base_class():
 
     # Run epf damage
     epf_dmg = EpfDamage(client)
+
     epf_dmg.load_remote_input_dataset("epfs", epf_dataset_id)
 
-    result_name = "earthquake_epf_dmg_result"
-    epf_dmg.set_parameter("result_name", result_name)
+    epf_dmg.set_parameter("result_name", "earthquake_epf_dmg_result")
     epf_dmg.set_parameter("mapping_id", mapping_id)
     epf_dmg.set_parameter("hazard_type", hazard_type)
     epf_dmg.set_parameter("hazard_id", hazard_id)
@@ -41,8 +41,7 @@ def run_with_base_class():
     epf_dmg = EpfDamage(client)
     epf_dmg.load_remote_input_dataset("epfs", epf_dataset_id)
 
-    result_name = "tsunami_epf_dmg_result"
-    epf_dmg.set_parameter("result_name", result_name)
+    epf_dmg.set_parameter("result_name", "tsunami_epf_dmg_result")
     epf_dmg.set_parameter("mapping_id", mapping_id)
     epf_dmg.set_parameter("hazard_type", hazard_type)
     epf_dmg.set_parameter("hazard_id", hazard_id)